FIx duplicate logging
