+++ .travis.yml
@@ -2,4 +2,4 @@ language: java
 jdk:
  - openjdk11
 script:
-  mvn test -fae -X
+  mvn test -fae
