+++ webdir-generator/src/main/java/com/kttdevelopment/webdir/generator/Vars.java
@@ -18,8 +18,6 @@ public abstract class Vars {
         public static boolean clear = false;
         public static boolean server = false;
 
-        public static boolean disableLogger = false;
-
         private static final AtomicInteger port = new AtomicInteger(40000);
 
         public synchronized static int assignPort(){
