+++ webdir-server/pom.xml
@@ -53,7 +53,7 @@
                 <configuration>
                     <useFile>false</useFile>
                     <trimStackTrace>false</trimStackTrace>
-                    <threadCount>1</threadCount>
+                    <reuseForks>false</reuseForks>
                 </configuration>
             </plugin>
         </plugins>
