+++ webdir-server/src/test/java/application/TestClear.java
@@ -1,3 +1,5 @@
+package application;
+
 import com.kttdevelopment.webdir.generator.Vars;
 import com.kttdevelopment.webdir.server.Main;
 import org.junit.Assert;
@@ -6,20 +8,11 @@ import org.junit.Test;
 import java.io.File;
 import java.io.IOException;
 
-public class ApplicationTests {
-
-    @Test
-    public void testSafeMode(){
-        Vars.Test.safemode = true;
-        Vars.Test.disableLogger = true;
-        Main.main(null);
-        Assert.assertTrue("Safe-mode should not load any plugins", Vars.Main.getPluginLoader().getPlugins().isEmpty());
-    }
+public class TestClear {
 
      @Test
     public void testClear() throws IOException{
         Vars.Test.safemode = true;
-        Vars.Test.clear = false;
 
         final File testRoot = new File(".root/testClear.html");
         final File testOutput = new File("_site/" + testRoot.getName());
@@ -37,7 +30,7 @@ public class ApplicationTests {
 
         Main.main(null);
         Assert.assertFalse("Generator did not remove file that was no longer present in root folder",testOutput.exists());
-        Vars.Test.clear = false;
     }
+
 
 }
