+++ webdir-server/src/test/java/server/render/TestPluginLoading.java
@@ -1,0 +1,35 @@
+package server.render;
+
+import com.kttdevelopment.webdir.generator.PluginLoader;
+import com.kttdevelopment.webdir.generator.Vars;
+import com.kttdevelopment.webdir.server.Main;
+import org.junit.Assert;
+import org.junit.Test;
+
+public class TestPluginLoading {
+
+    @Test
+    public void testExtendedPluginLoading(){
+        Vars.Test.server = true;
+
+        final String[] goodPlugins = {
+            "RenderTests",
+            "DuplicateRenderTests",
+            "ExchangeRenderTests",
+            "ExchangeDuplicateRenderTests",
+            "FileHandlerRenderTests",
+            "FileHandlerDuplicateRenderTests"
+        };
+
+        Vars.Test.assignPort();
+        Main.main(null);
+
+        final PluginLoader pluginLoader = Vars.Main.getPluginLoader();
+
+        for(final String goodPlugin : goodPlugins)
+            Assert.assertNotNull("Server should have loaded plugin: " + goodPlugin, pluginLoader.getPlugin(goodPlugin));
+
+        Assert.assertEquals("Server should have only loaded: " + goodPlugins.length + " plugins",goodPlugins.length,pluginLoader.getPlugins().size());
+    }
+
+}
