+++ webdir-server/src/test/java/server/throttler/TestThrottlerN1.java
@@ -1,4 +1,4 @@
-package server;
+package server.throttler;
 
 import com.kttdevelopment.webdir.generator.Vars;
 import com.kttdevelopment.webdir.server.Main;
@@ -12,36 +12,7 @@ import java.net.URI;
 import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-public class ThrottlerTests {
-
-    @Test
-    public void test0(){
-        Vars.Test.safemode = true;
-        Vars.Test.server = true;
-
-        final int conn = 0;
-        final String perm =
-            "groups:\n" +
-            "  default:\n" +
-            "    options:\n" +
-            "      default: true\n" +
-            "      connection-limit: " + conn;
-
-        TestFile.createTestFile(new File("permissions.yml").getAbsoluteFile(),perm);
-        TestFile.createTestFile(new File(".root/test.html"),"");
-
-        final int port = Vars.Test.assignPort();
-
-        Vars.Test.disableLogger = true;
-        Main.main(null);
-
-        final String url = "http://localhost:" + port;
-
-        try{
-            TestResponse.getResponseContent(URI.create(url + "/test"));
-            Assert.fail("Server should not have returned response for a connection limit of " + conn);
-        }catch(final ExecutionException | InterruptedException ignored){ }
-    }
+public class TestThrottlerN1 {
 
     @Test
     public void testN1() throws InterruptedException, BrokenBarrierException{
@@ -56,12 +27,10 @@ public class ThrottlerTests {
             "      default: true\n" +
             "      connection-limit: " + conn;
 
-        TestFile.createTestFile(new File("permissions.yml").getAbsoluteFile(),perm);
+        TestFile.createTestFile(new File("permissions.yml").getAbsoluteFile(), perm);
         TestFile.createTestFile(new File(".root/test.html"),"");
 
         final int port = Vars.Test.assignPort();
-
-        Vars.Test.disableLogger = true;
         Main.main(null);
 
         final String url = "http://localhost:" + port;
